Revert "Test/mail"
