+++ aea/mail/base.py
@@ -50,6 +50,12 @@ class Envelope:
         self._sender = sender
         self._protocol_id = protocol_id
         self._message = message
+        assert type(self._to) == str or self._to is None
+        try:
+            if self._to is not None and type(self._to) == str:
+                self._to.encode('utf-8')
+        except Exception:
+            assert False
 
     @property
     def to(self) -> Address:
@@ -131,8 +137,7 @@ class Envelope:
         protocol_id = envelope_pb.protocol_id
         message = envelope_pb.message
 
-        envelope = Envelope(to=to, sender=sender,
-                            protocol_id=protocol_id, message=message)
+        envelope = Envelope(to=to, sender=sender, protocol_id=protocol_id, message=message)
         return envelope
 
 
@@ -150,7 +155,7 @@ class InBox(object):
 
     def empty(self) -> bool:
         """
-        Check for a envelope on the in queue.
+        Check for a message on the in queue.
 
         :return: boolean indicating whether there is a message or not
         """
@@ -158,28 +163,28 @@ class InBox(object):
 
     def get(self, block: bool = True, timeout: Optional[float] = None) -> Envelope:
         """
-        Check for a envelope on the in queue.
+        Check for a message on the in queue.
 
         :param block: if true makes it blocking.
         :param timeout: times out the block after timeout seconds.
 
-        :return: the envelope object.
+        :return: the message object.
         :raises Empty: if the attempt to get a message fails.
         """
         logger.debug("Checks for message from the in queue...")
-        envelope = self._queue.get(block=block, timeout=timeout)
+        msg = self._queue.get(block=block, timeout=timeout)
         logger.debug("Incoming message: to='{}' sender='{}' protocol_id='{}' message='{}'"
-                     .format(envelope.to, envelope.sender, envelope.protocol_id, envelope.message))
-        return envelope
+                     .format(msg.to, msg.sender, msg.protocol_id, msg.message))
+        return msg
 
     def get_nowait(self) -> Optional[Envelope]:
         """
-        Check for a envelope on the in queue and wait for no time.
+        Check for a message on the in queue and wait for no time.
 
-        :return: the envelope object
+        :return: the message object
         """
-        envelope = self._queue.get_nowait()
-        return envelope
+        item = self._queue.get_nowait()
+        return item
 
 
 class OutBox(object):
@@ -196,22 +201,22 @@ class OutBox(object):
 
     def empty(self) -> bool:
         """
-        Check for a envelope on the out queue.
+        Check for a message on the out queue.
 
-        :return: boolean indicating whether there is a envelope or not
+        :return: boolean indicating whether there is a message or not
         """
         return self._queue.empty()
 
-    def put(self, envelope: Envelope) -> None:
+    def put(self, item: Envelope) -> None:
         """
-        Put an envelope into the queue.
+        Put an item into the queue.
 
-        :param envelope: the envelope.
+        :param item: the message.
         :return: None
         """
-        logger.debug("Put an envelope in the queue: to='{}' sender='{}' protocol_id='{}' message='{}'..."
-                     .format(envelope.to, envelope.sender, envelope.protocol_id, envelope.message))
-        self._queue.put(envelope)
+        logger.debug("Put a message in the queue: to='{}' sender='{}' protocol_id='{}' message='{}'..."
+                     .format(item.to, item.sender, item.protocol_id, item.message))
+        self._queue.put(item)
 
     def put_message(self, to: Address, sender: Address,
                     protocol_id: ProtocolId, message: bytes) -> None:
@@ -224,8 +229,7 @@ class OutBox(object):
         :param message: the content of the message.
         :return: None
         """
-        envelope = Envelope(to=to, sender=sender,
-                            protocol_id=protocol_id, message=message)
+        envelope = Envelope(to=to, sender=sender, protocol_id=protocol_id, message=message)
         self._queue.put(envelope)
