--- aea/skills/scaffold/skill.yaml
@@ -1,14 +1,0 @@
-name: ''
-authors: Fetch.AI Limited
-version: 0.1.0
-license: Apache 2.0
-url: ""
-behaviours:
-  - behaviour:
-    class_name: MyScaffoldBehaviour
-handler:
-  class_name: MyScaffoldHandler
-tasks:
-  - task:
-    class_name: MyScaffoldTask
-protocol: ''
