--- docs/version.md
@@ -1,1 +1,0 @@
-Current version of the Autonomous Econonmic Agent application is `0.1`.
