+++ tox.ini
@@ -31,8 +31,6 @@ commands = mypy aea examples tests scripts
 description = Build the documentation.
 basepython = python3.7
 deps = mkdocs
-       mkdocs-material
-       pymdown-extensions
 commands = mkdocs build --clean
 
 [testenv:docs-serve]
